+++ app/src/main/java/eu/homeanthill/di/KoinModules.kt
@@ -55,7 +55,7 @@ val viewModelModule = module {
   }
   viewModel { HomesListViewModel(homesRepository = get()) }
   viewModel { RoomsViewModel(homesRepository = get()) }
-  viewModel { DevicesListViewModel(devicesRepository = get(), homesRepository = get()) }
+  viewModel { DevicesListViewModel(devicesRepository = get(), homesRepository = get(), onlineRepository = get()) }
   viewModel { SensorFeatureValuesViewModel() }
   viewModel { ControllerFeatureValuesViewModel(devicesRepository = get()) }
   viewModel { OnlineFeatureValuesViewModel(onlineRepository = get(), devicesRepository = get()) }